enc(add): 2024-31-05

+++ app/src/main/java/com/sendiko/ola/database/BmiData.kt
@@ -12,6 +12,6 @@ data class BmiData(
     val tanggal: String,
     val beratBadan: Int,
     val tinggiBadan: Int,
-    val nilaiBmi: Int,
+    val nilaiBmi: Double,
     val gender: String,
 )
